cleanup bin heirarchy
